refactor: State 중복코드 Finished, Running으로 한 단계 추상화
